backend/requirements.txt
amqp
bcrypt
billiard
blinker
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
cors
cryptography
filelock
Flask
Flask-Cors
future
gevent
greenlet
idna
itsdangerous
Jinja2
kombu
MarkupSafe
prompt_toolkit
pycparser
PyJWT
PySocks
python-dateutil
redis
requests
requests-file
six
tldextract
tzdata
urllib3
vine
wcwidth
Werkzeug
zope.event
zope.interface
Werkzeug
python-dotenv
geopandas
numpy
shapely
scipy
pandas
tqdm
# New dependencies for the smart road data merger system
google-adk
tensorflow
sklearn
matplotlib
folium
geopy
news-api-client
fake-useragent